+++ packages/kit/types/private.d.ts
@@ -2,7 +2,7 @@
 // but which cannot be imported from `@sveltejs/kit`. Care should
 // be taken to avoid breaking changes when editing this file
 
-import { SSRNodeLoader, SSRRoute, ValidatedConfig } from './internal';
+import { ValidatedConfig } from './internal';
 
 export interface AdapterEntry {
 	/**
@@ -29,69 +29,6 @@ export interface AdapterEntry {
 }
 
 export type Body = JSONValue | Uint8Array | ReadableStream | import('stream').Readable;
-
-export interface Builder {
-	log: Logger;
-	rimraf(dir: string): void;
-	mkdirp(dir: string): void;
-
-	config: ValidatedConfig;
-	prerendered: Prerendered;
-
-	/**
-	 * Create entry points that map to individual functions
-	 * @param fn A function that groups a set of routes into an entry point
-	 */
-	createEntries(fn: (route: RouteDefinition) => AdapterEntry): void;
-
-	generateManifest: (opts: { relativePath: string; format?: 'esm' | 'cjs' }) => string;
-
-	getBuildDirectory(name: string): string;
-	getClientDirectory(): string;
-	getServerDirectory(): string;
-	getStaticDirectory(): string;
-
-	/**
-	 * @param dest the destination folder to which files should be copied
-	 * @returns an array of paths corresponding to the files that have been created by the copy
-	 */
-	writeClient(dest: string): string[];
-	/**
-	 *
-	 * @param dest
-	 */
-	writePrerendered(
-		dest: string,
-		opts?: {
-			fallback?: string;
-		}
-	): string[];
-	/**
-	 * @param dest the destination folder to which files should be copied
-	 * @returns an array of paths corresponding to the files that have been created by the copy
-	 */
-	writeServer(dest: string): string[];
-	/**
-	 * @param dest the destination folder to which files should be copied
-	 * @returns an array of paths corresponding to the files that have been created by the copy
-	 */
-	writeStatic(dest: string): string[];
-	/**
-	 * @param from the source file or folder
-	 * @param to the destination file or folder
-	 * @param opts.filter a function to determine whether a file or folder should be copied
-	 * @param opts.replace a map of strings to replace
-	 * @returns an array of paths corresponding to the files that have been created by the copy
-	 */
-	copy(
-		from: string,
-		to: string,
-		opts?: {
-			filter?: (basename: string) => boolean;
-			replace?: Record<string, string>;
-		}
-	): string[];
-}
 
 // Based on https://github.com/josh-hemphill/csp-typed-directives/blob/latest/src/csp.types.ts
 //
